Refactor code to extract shared logic
