+++ src/Commands/HydeRebuildStaticSiteCommand.php
@@ -3,6 +3,7 @@
 namespace Hyde\Framework\Commands;
 
 use Exception;
+use Hyde\Framework\Commands\Traits\TransfersMediaAssetsForBuildCommands;
 use Hyde\Framework\Hyde;
 use Hyde\Framework\Services\BuildService;
 use Hyde\Framework\Services\RebuildService;
@@ -13,6 +14,8 @@ use LaravelZero\Framework\Commands\Command;
  */
 class HydeRebuildStaticSiteCommand extends Command
 {
+    use TransfersMediaAssetsForBuildCommands;
+
     /**
      * The signature of the command.
      *
@@ -52,7 +55,8 @@ class HydeRebuildStaticSiteCommand extends Command
         $time_start = microtime(true);
 
         if ($this->argument('path') === '_media') {
-            return $this->handleMediaFiles();
+            $this->transferMediaAssets();
+            return 0;
         }
 
         $this->path = $this->sanitizePathString($this->argument('path'));
@@ -105,8 +109,7 @@ class HydeRebuildStaticSiteCommand extends Command
      */
     public function validate(): void
     {
-        if (! (
-            str_starts_with($this->path, '_docs') ||
+        if (!(str_starts_with($this->path, '_docs') ||
             str_starts_with($this->path, '_posts') ||
             str_starts_with($this->path, '_pages') ||
             str_starts_with($this->path, 'resources/views/pages')
@@ -114,7 +117,7 @@ class HydeRebuildStaticSiteCommand extends Command
             throw new Exception("Path [$this->path] is not in a valid source directory.", 400);
         }
 
-        if (! file_exists(Hyde::path($this->path))) {
+        if (!file_exists(Hyde::path($this->path))) {
             throw new Exception("File [$this->path] not found.", 404);
         }
     }
@@ -131,38 +134,5 @@ class HydeRebuildStaticSiteCommand extends Command
         $this->warn($exception->getMessage());
 
         return $exception->getCode();
-    }
-
-    /**
-     * Handle the media files command.
-     *
-     * @return int
-     */
-    public function handleMediaFiles(): int
-    {
-        $collection = glob(Hyde::path('_media/*.{png,svg,jpg,jpeg,gif,ico,css,js}'), GLOB_BRACE);
-        $collection = array_merge($collection, [
-            Hyde::path('resources/frontend/hyde.css'),
-            Hyde::path('resources/frontend/hyde.js'),
-        ]);
-        if (sizeof($collection) < 1) {
-            $this->line('No Media Assets found. Skipping...');
-            $this->newLine();
-        } else {
-            $this->comment('Transferring Media Assets...');
-            $this->withProgressBar(
-                $collection,
-                function ($filepath) {
-                    if ($this->getOutput()->isVeryVerbose()) {
-                        $this->line(' > Copying media file '
-                            .basename($filepath).' to the output media directory');
-                    }
-                    copy($filepath, Hyde::path('_site/media/'.basename($filepath)));
-                }
-            );
-            $this->newLine(2);
-        }
-
-        return 1;
     }
 }
