+++ src/Commands/Traits/BuildActionRunner.php
@@ -1,0 +1,50 @@
+<?php
+
+namespace Hyde\Framework\Commands\Traits;
+
+use Hyde\Framework\Services\BuildService;
+use Hyde\Framework\Services\CollectionService;
+use Hyde\Framework\StaticPageBuilder;
+
+/** 
+ * Offloads logic for static site building commands.
+ * @see \Hyde\Framework\Commands\HydeBuildStaticSiteCommand
+ * @internal
+ */
+trait BuildActionRunner
+{
+    /** @internal */
+    protected function canRunBuildAction(array $collection, string $name, ?string $verb = null): bool
+    {
+        if (sizeof($collection) < 1) {
+            $this->line('No ' . $name . ' found. Skipping...');
+            $this->newLine();
+            return false;
+        }
+
+        $this->comment(($verb ?? 'Creating') . " $name...");
+        return true;
+    }
+
+    /** @internal */
+    protected function runBuildAction(string $model)
+    {
+        $collection = CollectionService::getSourceFileListForModel($model);
+        $modelName = $this->getModelPluralName($model);
+        if ($this->canRunBuildAction($collection, $modelName)) {
+            $this->withProgressBar(
+                $collection,
+                function ($basename) use ($model) {
+                    new StaticPageBuilder(
+                        BuildService::getParserInstanceForModel(
+                            $model,
+                            $basename
+                        )->get(),
+                        true
+                    );
+                }
+            );
+            $this->newLine(2);
+        }
+    }
+}
